search bar functionality
Search bar:
or the mean time I made it that when I pressed enter key, depending what you typed it will go to a "new screen" demo page that I have to test it works, with the name that you search for, I made it just so the enter functionality can work. And I think whenever we are ready to test it with the API then I can most likely implement it here.